requirements.txt
absl-py==2.2.2
accelerate==1.5.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.1
asgiref==3.8.1
attr==0.3.1
attrs==25.3.0
Automat==22.10.0
azure-core==1.32.0
azure-storage-blob==12.25.1
Babel==2.10.3
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
black==25.1.0
bleach==5.0.1
blinker==1.7.0
blis==1.2.1
boto==2.49.0
boto3==1.37.24
botocore==1.37.24
bottle==0.12.25
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
catalogue==2.0.10
cattrs==22.2.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.22
cleo==2.1.0
click==8.1.6
cloud-init==24.4
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
configobj==5.0.8
constantly==23.10.4
contourpy==1.3.1
crashtest==0.4.1
crfm-helm==0.5.5
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
dacite==1.9.2
dataclasses-json==0.6.7
datamodel-code-generator==0.26.1
datasets==2.21.0
dbus-python==1.3.2
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.32.2
dill==0.3.8
distlib==0.3.8
distro==1.9.0
distro-info==1.7+build1
Django==5.1.7
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.2.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dnspython==2.7.0
docling==2.28.0
docling-core==2.23.3
docling-ibm-models==3.4.1
docling-parse==4.0.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
dulwich==0.22.8
durationpy==0.9
easyocr==1.7.2
edge-tts==7.0.0
email_validator==2.2.0
et_xmlfile==2.0.0
expiringdict==1.2.2
faiss-cpu==1.10.0
Faker==37.1.0
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.18.0
filetype==1.2.0
findpython==0.6.3
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
genson==1.3.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.11.4
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
gyp==0.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.20.4
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.26.5
humanfriendly==10.0
humansignal-drf-yasg==1.21.10.post1
hyperlink==21.0.0
idna==3.10
ijson==3.3.0
imageio==2.37.0
importlib-resources==5.13.0
importlib_metadata==8.6.1
incremental==22.10.0
inflect==5.6.2
inflection==0.5.1
installer==0.7.0
isodate==0.7.2
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.2
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==2.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
kiwisolver==1.4.8
kubernetes==32.0.1
label-studio==1.16.0
label-studio-sdk==1.0.10
labelImg==1.8.6
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-huggingface==0.1.2
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langcodes==3.5.0
langsmith==0.2.11
language_data==1.3.0
latex2mathml==3.77.0
launchdarkly-server-sdk==8.2.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lazy_loader==0.4
lockfile==0.12.2
lxml==5.3.1
lxml_html_clean==0.4.1
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
marko==2.1.2
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.8.2
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==10.6.0
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.31.0
netifaces==0.11.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.21.0
openai==1.68.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
ordered-set==4.0.2
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pathspec==0.12.1
pbs-installer==2025.3.17
pillow==11.1.0
pipenv==2023.12.1
pkginfo==1.12.1.2
platformdirs==4.2.0
pluggy==1.5.0
poetry==2.1.1
poetry-core==2.1.1
posthog==3.21.0
preshed==3.0.9
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
pulsar-client==3.6.1
pyarrow==19.0.1
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.13
pyboxen==1.3.0
pyclipper==1.3.0.post6
pycparser==2.22
pycurl==7.45.3
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.17.2
PyGObject==3.48.2
PyHamcrest==2.1.0
pyhocon==0.3.61
PyJWT==2.7.0
pylatexenc==2.10
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
pyRFC3339==2.0.1
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.4
python-pptx==1.0.2
pytz==2022.7.1
PyYAML==6.0.2
RapidFuzz==3.12.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.7.1
rouge_score==0.1.2
rpds-py==0.23.1
rq==1.16.2
rsa==4.9
rstr==3.2.2
rtree==1.4.0
rules==3.4
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.0
scipy==1.15.2
SecretStorage==3.3.3
semchunk==2.2.2
semver==3.0.4
sentence-transformers==3.3.1
sentry-sdk==2.25.0
service-identity==24.1.0
shapely==2.0.7
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.39
sqlitedict==2.1.0
sqlparse==0.5.3
srsly==2.5.1
srt==3.5.3
starlette==0.46.1
streamlit==1.43.2
streamlit-pdf-viewer==0.0.21
sympy==1.13.1
systemd-python==235
tabulate==0.9.0
tenacity==9.0.0
thinc==8.3.4
threadpoolctl==3.6.0
tifffile==2025.3.13
tiktoken==0.9.0
tk==0.1.0
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.2
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
triton==3.2.0
trove-classifiers==2025.3.19.19
Twisted==24.3.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ubuntu-pro-client==8001
ujson==5.10.0
unattended-upgrades==0.1
uncertainty-calibration==0.1.4
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.3
wadllib==1.3.6
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.4
weasel==0.4.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
whisper==1.1.10
wrapt==1.17.2
XlsxWriter==3.2.2
xmljson==0.2.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zope.interface==6.1
zstandard==0.18.0
